requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==4.1
django-extensions==3.2.0
django-filter==22.1
djangorestframework==3.13.1
pycodestyle==2.9.1
pytz==2022.2.1
sqlparse==0.4.2
toml==0.10.2
